\name{aesExams-package}
\alias{aesExams-package}
\alias{aesExams}
\docType{package}
\title{
\packageTitle{aesExams}
}
\description{
\packageDescription{aesExams}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{aesExams}
\packageIndices{aesExams}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{aesExams}

Maintainer: \packageMaintainer{aesExams}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
