% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradeProcessing.R
\name{createGradeFeedback}
\alias{createGradeFeedback}
\title{Create Grade Feedback}
\usage{
createGradeFeedback(s, g, examname = "ExamIB")
}
\arguments{
\item{s}{A list of student objects, each containing detailed information including answers to exam questions.}

\item{g}{A list of corresponding grades for each question for the students.}

\item{examname}{A string indicating the name of the exam, default is "ExamIB".}
}
\value{
A list where each element is a vector of strings composing the feedback for a single student. This feedback is not currently saved to a file within the function and might need to be captured or processed further by the calling code.
}
\description{
Generates feedback for each student based on their answers and the corresponding grades for an exam. The feedback includes a summary of the student's performance by question, with specifics on their answers and the correct answers.
}
\examples{
students <- list(student1, student2)
grades <- list(grades1, grades2)
feedback <- createGradeFeedback(students, grades, "Final Exam")
}
