% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradeProcessing.R
\name{createGradeUpload}
\alias{createGradeUpload}
\title{Create Grade Upload File}
\usage{
createGradeUpload(s, g, fn = "gradeUpload.csv")
}
\arguments{
\item{s}{A list of student objects, each containing student information including the student ID and username.}

\item{g}{A list of corresponding grades for the students.}

\item{fn}{The filename for the generated CSV file, default is "gradeUpload.csv".}
}
\value{
None, but creates or overwrites a CSV file with the name specified by `fn`, containing the grades ready for upload.
}
\description{
Generates a CSV file for uploading grades, containing student IDs, usernames, and their respective grades.
}
\examples{
students <- list(student1, student2)
grades <- list(grade1, grade2)
createGradeUpload(students, grades, "grades.csv")
}
