% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradeQuestionTypes.R
\name{grade.mchoice}
\alias{grade.mchoice}
\title{Grade Multiple Choice Responses}
\usage{
grade.mchoice(correct, answer, allmatch = TRUE)
}
\arguments{
\item{correct}{A list containing the correct answers' metadata, including the solutions.}

\item{answer}{A list containing the student's answers, specifically whether each option was selected.}

\item{allmatch}{Logical indicating whether to match all answers exactly. Default is TRUE.}
}
\value{
A numeric value representing the proportion of correct answers out of the total possible correct answers.
}
\description{
This function calculates the proportion of correctly matched answers for multiple choice questions.
It compares a vector of correct answers to a vector of student's answers, allowing for different lengths
of answer vectors and adjusting them accordingly.
}
\examples{
# Assuming correct and answer are predefined lists with the necessary structure
grade.mchoice(correct, answer)

}
