% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string2num}
\alias{string2num}
\title{Convert String Representations of Numbers to Numeric Format}
\usage{
string2num(x = "")
}
\arguments{
\item{x}{A character vector where each element is a string representing a number.
The default is an empty string, which will result in an NA value.}
}
\value{
A numeric vector with the same length as `x`, where each string representation
  of a number has been converted to numeric format. Elements that cannot be converted
  to a number return NA.
}
\description{
This function takes a vector of strings representing numbers in various formats
(e.g., with letters, commas, as fractions, or percentages) and attempts to convert
them into numeric format. It handles fractions and percentages specifically, converting
them to their decimal equivalents. Other non-numeric characters are removed, and the
remaining numeric string is converted to a numeric type. Warnings are suppressed during
the conversion process. NA is returned for elements that cannot be converted.
}
\examples{
string2num(c("10\%", "1/4", "3,000", "5 apples", "2.5"))
# Expected output: c(0.1, 0.25, 3000, NA, 2.5)

}
