% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_d2l_answers.R
\name{studentAns}
\alias{studentAns}
\title{Student Answers Processing Function}
\usage{
studentAns(
  path_to_directory,
  path_to_exam,
  ansfn = "Exam I Part B (there is also a part A) - Attempt Details.csv",
  idmapfn = "../../2024Spring Genetics (BIOL-305-01)_GradesExport_2024-02-22-01-07.csv",
  examName = "exam1Answers",
  solnName = "exam1Answers-solutions",
  answersFN = "exam1SpecificAnswers.RDS"
)
}
\arguments{
\item{path_to_directory}{The directory path where the grades are stored.}

\item{path_to_exam}{The directory path where the exam was created.}

\item{ansfn}{Filename of the answer file from D2L, defaulting to "Exam I Part B (there is also a part A) - Attempt Details.csv".}

\item{idmapfn}{Filename for the ID mapping file, defaulting to a relative path for the 2024 Spring Genetics course grades export.}

\item{examName}{The base name for exam answers files.}

\item{solnName}{The base name for solution files.}

\item{answersFN}{Filename for the specific answers file in RDS format, default "exam1SpecificAnswers.RDS".}
}
\value{
A list where each element corresponds to a student, including their answers,
        student information, and the correct exam answers.
}
\description{
This function processes exam answers from students, associating each student's
responses with their ID and the specific questions from the exam. It generates
a list of student answers, the questions, and the correct answers from a provided exam.
}
\examples{
studentAns("path/to/grades", "path/to/exam")

}
